Feat. Train: 모델 학습 구조 변경

+++ Model/utils/matrix.py
@@ -61,38 +61,38 @@ class Matrix:
         for metric in self._list.values():
             metric.reset_states()
 
-    def get_accuracy(self) -> np.float32:
+    def get_accuracy(self) -> float:
         mat = self.get_confusion_matrix()
         acc = (mat.tp + mat.tn) / (mat.tp + mat.tn + mat.fp + mat.fn) if (mat.tp + mat.tn + mat.fp + mat.fn) != 0 else 0
         return acc
 
-    def get_recall(self) -> np.float32:
+    def get_recall(self) -> float:
         mat = self.get_confusion_matrix()
         recall = mat.tp / (mat.tp + mat.fn) if (mat.tp + mat.fn) != 0 else 0
         return recall
 
-    def get_ok_recall(self) -> np.float32:
+    def get_ok_recall(self) -> float:
         mat = self.get_confusion_matrix()
         ok_recall = mat.tn / (mat.tn + mat.fp) if (mat.tn + mat.fp) != 0 else 0
         return ok_recall
 
-    def get_precision(self) -> np.float32:
+    def get_precision(self) -> float:
         mat = self.get_confusion_matrix()
         precision = mat.tp / (mat.tp + mat.fp) if (mat.tp + mat.fp) != 0 else 0
         return precision
 
-    def get_ok_precision(self) -> np.float32:
+    def get_ok_precision(self) -> float:
         mat = self.get_confusion_matrix()
         ok_precision = mat.tn / (mat.tn + mat.fn) if (mat.tn + mat.fn) != 0 else 0
         return ok_precision
 
-    def get_f1_score(self) -> np.float64:
+    def get_f1_score(self) -> float:
         recall = self.get_recall()
         precision = self.get_precision()
         f1 = (2 * precision * recall) / (precision + recall) if (precision + recall) != 0 else 0
         return f1
 
-    def get_ok_f1_score(self) -> np.float64:
+    def get_ok_f1_score(self) -> float:
         ok_recall = self.get_ok_recall()
         ok_precision = self.get_ok_precision()
         ok_f1 = (2 * ok_precision * ok_recall) / (ok_precision + ok_recall) if (ok_precision + ok_recall) != 0 else 0
